+++ features/Context/Page/Base/Base.php
@@ -122,8 +122,27 @@ class Base extends Page
      * Overriden for compatibility with links
      *
      * @param string $locator
+     *
+     * @throws ElementNotFoundException
      */
     public function pressButton($locator)
+    {
+        $button = $this->getButton($locator);
+        if (null === $button) {
+            throw new ElementNotFoundException($this->getSession(), 'button', 'id|name|title|alt|value', $locator);
+        }
+
+        $button->click();
+    }
+
+    /**
+     * Get button
+     *
+     * @param string $locator
+     *
+     * @return NodeElement
+     */
+    public function getButton($locator)
     {
         // Search with exact name at first
         $button = $this->find('xpath', sprintf("//button[text() = '%s']", $locator));
@@ -147,11 +166,7 @@ class Base extends Page
             );
         }
 
-        if (null === $button) {
-            throw new ElementNotFoundException($this->getSession(), 'button', 'id|name|title|alt|value', $locator);
-        }
-
-        $button->click();
+        return $button;
     }
 
     /**