+++ src/Pim/Bundle/UserBundle/Context/UserContext.php
@@ -4,6 +4,7 @@ namespace Pim\Bundle\UserBundle\Context;
 
 use Symfony\Component\Security\Core\SecurityContextInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Oro\Bundle\UserBundle\Entity\User;
 use Oro\Bundle\SecurityBundle\SecurityFacade;
 use Pim\Bundle\CatalogBundle\Manager\LocaleManager;
 use Pim\Bundle\CatalogBundle\Manager\ChannelManager;
@@ -292,14 +293,18 @@ class UserContext
     /**
      * Get authenticated user
      *
-     * @return UserInterface|string|null
+     * @return User|null
      */
     public function getUser()
     {
-        $token = $this->securityContext->getToken();
-
-        if ($token !== null) {
-            return $token->getUser();
+        if (null === $token = $this->securityContext->getToken()) {
+            return null;
         }
+
+        if (!is_object($user = $token->getUser())) {
+            return null;
+        }
+
+        return $user;
     }
 }